Remove host header from listener rules.
Removed the host header from both author and author-api load balancer listener rules.
This change was to support the author short url in addition to the regular service address.
In doing this change I had to increase the priority of the listener rule for author to ensure that the path wildcard does not route traffic intended for other services, namely validator and publisher.